delete confusing extension methods
