+++ core-tests/shared/src/test/scala/zio/prelude/PartialOrdSpec.scala
@@ -13,18 +13,18 @@ object PartialOrdSpec extends ZIOBaseSpec {
         test("set")(checkAllLaws(PartialOrdLaws)(Gen.setOf(Gen.int)))
       ),
       test("map compareSoft") {
-        assert(Map.empty[Int, Int] compareSoft Map.empty[Int, Int])(equalTo(Ordering.Equals)) &&
-        assert(Map(1 -> 2) compareSoft Map.empty[Int, Int])(equalTo(Ordering.GreaterThan)) &&
-        assert(Map.empty[Int, Int] compareSoft Map(1 -> 2))(equalTo(Ordering.LessThan)) &&
-        assert(Map(1 -> 2) compareSoft Map(1 -> 2))(equalTo(Ordering.Equals)) &&
-        assert(Map(1 -> 2) compareSoft Map(1 -> 3))(equalTo(Ordering.LessThan)) &&
-        assert(Map(1 -> 3) compareSoft Map(1 -> 2))(equalTo(Ordering.GreaterThan)) &&
-        assert(Map(1 -> 2) compareSoft Map(1 -> 3, 2 -> 4))(equalTo(Ordering.LessThan)) &&
-        assert(Map(1 -> 3, 2 -> 4) compareSoft Map(1 -> 2))(equalTo(Ordering.GreaterThan)) &&
-        assert(Map(1 -> 2) compareSoft Map(2 -> 2))(equalTo(PartialOrdering.Incomparable)) &&
-        assert(Map(1 -> 2, 3 -> 3) compareSoft Map(3 -> 3, 2 -> 2))(equalTo(PartialOrdering.Incomparable)) &&
-        assert(Map(1 -> 2, 3 -> 3) compareSoft Map(3 -> 1, 2 -> 2))(equalTo(PartialOrdering.Incomparable)) &&
-        assert(Map("William" -> "Will") compareSoft Map("William" -> "Bill"))(equalTo(Ordering.GreaterThan))
+        assert(PartialOrd.compareSoft(Map.empty[Int, Int], Map.empty[Int, Int]))(equalTo(Ordering.Equals)) &&
+        assert(PartialOrd.compareSoft(Map(1 -> 2), Map.empty[Int, Int]))(equalTo(Ordering.GreaterThan)) &&
+        assert(PartialOrd.compareSoft(Map.empty[Int, Int], Map(1 -> 2)))(equalTo(Ordering.LessThan)) &&
+        assert(PartialOrd.compareSoft(Map(1 -> 2), Map(1 -> 2)))(equalTo(Ordering.Equals)) &&
+        assert(PartialOrd.compareSoft(Map(1 -> 2), Map(1 -> 3)))(equalTo(Ordering.LessThan)) &&
+        assert(PartialOrd.compareSoft(Map(1 -> 3), Map(1 -> 2)))(equalTo(Ordering.GreaterThan)) &&
+        assert(PartialOrd.compareSoft(Map(1 -> 2), Map(1 -> 3, 2 -> 4)))(equalTo(Ordering.LessThan)) &&
+        assert(PartialOrd.compareSoft(Map(1 -> 3, 2 -> 4), Map(1 -> 2)))(equalTo(Ordering.GreaterThan)) &&
+        assert(PartialOrd.compareSoft(Map(1 -> 2), Map(2 -> 2)))(equalTo(PartialOrdering.Incomparable)) &&
+        assert(PartialOrd.compareSoft(Map(1 -> 2, 3 -> 3), Map(3 -> 3, 2 -> 2)))(equalTo(PartialOrdering.Incomparable)) &&
+        assert(PartialOrd.compareSoft(Map(1 -> 2, 3 -> 3), Map(3 -> 1, 2 -> 2)))(equalTo(PartialOrdering.Incomparable)) &&
+        assert(PartialOrd.compareSoft(Map("William" -> "Will"), Map("William" -> "Bill")))(equalTo(Ordering.GreaterThan))
       }
     )
 }
